KEYCLOAK-14817 Allow JS adapter to be bundled as ES module
