+++ .github/dependabot.yml
@@ -1,0 +1,6 @@
+version: 2
+updates:
+  - package-ecosystem: npm
+    directory: "/adapters/oidc/js"
+    schedule:
+      interval: weekly
