+++ adapters/oidc/js/.browserslistrc
@@ -1,0 +1,6 @@
+IE 11
+last 2 Chrome versions
+last 2 Firefox versions
+last 2 Safari versions
+Firefox ESR
+not dead
