+++ adapters/oidc/js/.gitignore
@@ -1,0 +1,4 @@
+node/
+node_modules
+packages/*/es/
+packages/*/lib/
