+++ adapters/oidc/js/package.json
@@ -1,0 +1,29 @@
+{
+  "private": true,
+  "workspaces": [
+    "packages/keycloak-js",
+    "packages/keycloak-authz"
+  ],
+  "engines": {
+    "node": "^16"
+  },
+  "scripts": {
+    "build": "npm run build --workspaces",
+    "bundle": "rimraf target/classes && webpack"
+  },
+  "devDependencies": {
+    "@babel/core": "^7.13.16",
+    "@babel/preset-env": "^7.13.5",
+    "@types/node": "^14.14.41",
+    "@types/terser-webpack-plugin": "^5.0.2",
+    "babel-loader": "^8.2.2",
+    "copyfiles": "^2.4.1",
+    "promise-polyfill": "^8.2.0",
+    "rimraf": "^3.0.2",
+    "terser-webpack-plugin": "^5.1.1",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.2.2",
+    "webpack": "^5.35.1",
+    "webpack-cli": "^4.5.0"
+  }
+}
