+++ adapters/oidc/js/packages/keycloak-authz/tsconfig.es.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "module": "ES2015",
+    "outDir": "./es"
+  }
+}
