--- distribution/adapters/js-adapter-npm-zip/assembly.xml
@@ -1,45 +1,0 @@
-<!--
-  ~ Copyright 2016 Red Hat, Inc. and/or its affiliates
-  ~ and other contributors as indicated by the @author tags.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~ http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<assembly>
-    <id>war-dist</id>
-
-    <formats>
-        <format>zip</format>
-    </formats>
-    <includeBaseDirectory>true</includeBaseDirectory>
-
-    <files>
-        <file>
-            <source>src/main/resources/package.json</source>
-            <outputDirectory>/</outputDirectory>
-            <filtered>true</filtered>
-        </file>
-    </files>
-
-    <fileSets>
-        <fileSet>
-            <directory>${project.build.directory}/unpacked/js-adapter</directory>
-            <outputDirectory>dist/</outputDirectory>
-            <includes>
-                <include>**/*.js</include>
-                <include>**/*.map</include>
-                <include>**/*.d.ts</include>
-            </includes>
-        </fileSet>
-    </fileSets>
-</assembly>
