--- distribution/adapters/js-adapter-npm-zip/src/main/resources/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "keycloak-js",
-  "version": "${project.version}",
-  "description": "Keycloak Adapter",
-  "main": "dist/keycloak.js",
-  "typings": "dist/keycloak.d.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/keycloak/keycloak"
-  },
-  "author": "Keycloak",
-  "license": "Apache-2.0",
-  "homepage": "https://www.keycloak.org",
-  "keywords": [
-    "keycloak",
-    "sso",
-    "oauth",
-    "oauth2",
-    "authentication"
-  ],
-  "dependencies": {
-    "base64-js": "1.3.1",
-    "js-sha256": "0.9.0"
-  }
-}
