+++ distribution/adapters/pom.xml
@@ -51,7 +51,6 @@
                 <module>jetty92-adapter-zip</module>
                 <module>jetty93-adapter-zip</module>
                 <module>jetty94-adapter-zip</module>
-                <module>js-adapter-npm-zip</module>
             </modules>
         </profile>
         <profile>
