+++ themes/pom.xml
@@ -72,8 +72,7 @@
                     </execution>
                 </executions>
                 <configuration>
-                    <nodeVersion>v12.13.0</nodeVersion>
-                    <npmVersion>6.9.0</npmVersion>
+                    <nodeVersion>v16.0.0</nodeVersion>
                     <installDirectory>${project.basedir}</installDirectory>
                 </configuration>
             </plugin>
