+++ themes/src/main/resources/theme/keycloak.v2/account/src/app/keycloak-service/keycloak.service.ts
@@ -14,10 +14,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import {KeycloakLoginOptions} from "../../../../../../../../../../adapters/oidc/js/src/main/resources/keycloak";
+import Keycloak, { KeycloakLoginOptions } from '../../../../../../../../../../adapters/oidc/js/packages/keycloak-js'
 
 declare const baseUrl: string;
-export type KeycloakClient = Keycloak.KeycloakInstance;
+export type KeycloakClient = Keycloak;
 
 export class KeycloakService {
     private keycloakAuth: KeycloakClient;
