+++ themes/src/main/resources/theme/keycloak.v2/account/src/tsconfig.json
@@ -4,6 +4,7 @@
     "module": "commonjs",
     "noEmit": true,
     "moduleResolution": "node",
+    "esModuleInterop": true,
     "sourceMap": true,
     "emitDecoratorMetadata": true,
     "experimentalDecorators": true,
@@ -15,8 +16,5 @@
   },
   "include": [
     "./app/**/*.ts?"
-  ],
-  "files": [
-    "../../../../../../../../adapters/oidc/js/src/main/resources/keycloak.d.ts"
   ]
 }
